rename vendor to _vendor
