+++ setup.cfg
@@ -12,7 +12,7 @@ universal=1
 [pylama]
 linters = mccabe,pep8,pyflakes
 ignore = D203,C901
-skip = .tox/*,.venv/*,nornir/vendor/*
+skip = .tox/*,.venv/*,nornir/_vendor/*
 
 [pylama:pep8]
 max_line_length = 100
@@ -58,5 +58,5 @@ ignore_errors = True
 [mypy-tests.*]
 ignore_errors = True
 
-[mypy-nornir.vendor.*]
+[mypy-nornir._vendor.*]
 ignore_errors = True
