+++ tests/core/test_inventory.py
@@ -3,7 +3,7 @@ import os
 from nornir.core import inventory
 from nornir.core.deserializer import inventory as deserializer
 
-from nornir.vendor.pydantic import ValidationError
+from nornir._vendor.pydantic import ValidationError
 
 import pytest
